+++ app/crvite/src/components/ChatHistory.tsx
@@ -3,7 +3,7 @@ import { handle_select_chat, refresh_chats } from '@/handlers/chats_handlers';
 import { Chat } from '@/types/types';
 
 import { use_chats } from '@/hooks/hooks';
-import { start_delete_chat } from '@/api/api';
+import { start_delete_chat } from '@/api/api_user_data';
 import { add_log } from '@/handlers/log_handlers';
 import { MessageCirclePlus, Trash2 } from 'lucide-react';
 import { set_messages } from '@/handlers/message_handlers';
