+++ app/crvite/src/components/Table_Files.tsx
@@ -3,7 +3,7 @@
 import { useRef } from 'react';
 
 import * as React from "react"
-import { fetch_db_files_metadata } from "@/api/api"
+import { fetch_db_files_metadata } from "@/api/api_files"
 import { use_current_collection, use_selected_files, use_files } from "@/hooks/hooks"
 import { set_selected_files } from "@/handlers/file_handlers"
 import { FileData } from "@/types/types"
